requirements.txt
streamlit==1.0.0
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.1



